packages/grafana-report-fetching/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "grafana-report-fetching"
version = "0.1.0"
description = "Installable Grafana-backed report fetching CLI"
requires-python = ">=3.10"
dependencies = [
    "PyYAML>=6.0",
    "cptools-common @ git+ssh://git@github.com/cpoopc/share-libs.git@main#subdirectory=tools/python/libs/common",
    "cptools-grafana @ git+ssh://git@github.com/cpoopc/share-libs.git@main#subdirectory=tools/python/libs/grafana",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
grafana-report-fetching = "cptools_grafana_report_fetching.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cptools_grafana_report_fetching*"]

[tool.setuptools.package-data]
cptools_grafana_report_fetching = [
    "assets/.env.example",
    "assets/profile-aliases.yaml",
    "assets/core-metrics-daily.yaml",
]

[tool.uv.sources]
cptools-common = { path = "../../tools/python/libs/common", editable = true }
cptools-grafana = { path = "../../tools/python/libs/grafana", editable = true }
